requirements.txt
Django==3.0.6
django-ckeditor==6.0.0
django-cleanup==5.0.0
django-js-asset==1.2.2
django-modeltranslation==0.16
Pillow==7.1.2
python-gettext==4.0
pytz==2020.1
setuptools==39.1.0
six==1.15.0
sqlparse==0.3.1
